backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-browser-backend"
version = "0.1.0"
description = "AI-powered browser agent backend — goal-based cognitive agent with LangGraph"
requires-python = ">=3.11"
dependencies = [
    # Agent Framework
    "langgraph>=0.4.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-ollama>=0.3.0",
    "langchain-groq>=0.3.0",

    # API Server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.0",

    # Data Validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",

    # Logging
    "structlog>=24.0.0",

    # Utilities
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
]
playwright = [
    "playwright>=1.49.0",
]
cli = [
    "rich>=13.0.0",
    "click>=8.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.hatch.build.targets.wheel]
packages = ["src/agent_core"]
